app/requirements.txt
boto3
clerk-sdk-python
fastapi
Flask
requests
uvicorn
chardet
modal
stripe
tenacity
svix
python-dotenv
python-multipart